move to pkg/epforwarder
